refactor(module): fcfs-issue -> fcfs-coupon-domain-fcfs rename
