+++ api/src/main/java/com/coupop/api/CouponController.java
@@ -5,9 +5,9 @@ import com.coupop.api.dto.HistoryResponse;
 import com.coupop.api.dto.IssuanceRequest;
 import com.coupop.api.dto.ResendRequest;
 import com.coupop.api.support.RequestTime;
-import com.coupop.fcfscoupon.domain.coupon.CouponService;
 import com.coupop.fcfscoupon.common.exception.ApiException;
-import com.coupop.fcfsissue.FcfsIssueService;
+import com.coupop.fcfscoupon.domain.coupon.CouponService;
+import com.coupop.fcfscoupon.domain.fcfs.FcfsIssueService;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.FieldError;
 import org.springframework.validation.annotation.Validated;
