+++ api/src/test/java/com/coupop/CouponAcceptanceTest.java
@@ -10,10 +10,10 @@ import static org.springframework.http.HttpStatus.NOT_FOUND;
 import static org.springframework.http.HttpStatus.OK;
 
 import com.coupop.api.dto.HistoryRequest;
+import com.coupop.api.dto.HistoryResponse;
 import com.coupop.api.dto.IssuanceRequest;
 import com.coupop.api.dto.ResendRequest;
-import com.coupop.api.dto.HistoryResponse;
-import com.coupop.fcfsissue.model.FcfsIssuePolicy;
+import com.coupop.fcfscoupon.domain.fcfs.model.FcfsIssuePolicy;
 import com.coupop.testconfig.IntegrationTestConfig;
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
