+++ api/src/test/java/com/coupop/api/CouponControllerTest.java
@@ -20,10 +20,10 @@ import com.coupop.api.support.RequestTime;
 import com.coupop.fcfscoupon.domain.coupon.CouponIssueHistoriesResponseConverter;
 import com.coupop.fcfscoupon.domain.coupon.CouponService;
 import com.coupop.fcfscoupon.domain.coupon.exception.HistoryNotFoundException;
-import com.coupop.fcfsissue.FcfsIssueService;
-import com.coupop.fcfsissue.exception.CouponNotOpenedException;
-import com.coupop.fcfsissue.exception.CouponOutOfStockException;
-import com.coupop.fcfsissue.exception.EmailAlreadyUsedException;
+import com.coupop.fcfscoupon.domain.fcfs.FcfsIssueService;
+import com.coupop.fcfscoupon.domain.fcfs.exception.CouponNotOpenedException;
+import com.coupop.fcfscoupon.domain.fcfs.exception.CouponOutOfStockException;
+import com.coupop.fcfscoupon.domain.fcfs.exception.EmailAlreadyUsedException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.time.LocalTime;
 import java.util.List;
