+++ fcfs-coupon-domain-fcfs/src/main/java/com/coupop/fcfscoupon/domain/fcfs/model/FcfsIssueRepository.java
@@ -1,0 +1,10 @@
+package com.coupop.fcfscoupon.domain.fcfs.model;
+
+public interface FcfsIssueRepository {
+
+    Long add(final String email);
+
+    Long remove(final String email);
+
+    Long getCount();
+}
