+++ fcfs-coupon-domain-fcfs/src/test/java/com/coupop/fcfscoupon/domain/TestContext.java
@@ -1,0 +1,7 @@
+package com.coupop.fcfscoupon.domain;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TestContext {
+}
