+++ fcfs-coupon-domain-fcfs/src/test/java/com/coupop/fcfscoupon/domain/fcfs/testconfig/DatabaseSetUp.java
@@ -1,4 +1,4 @@
-package com.coupop.testconfig;
+package com.coupop.fcfscoupon.domain.fcfs.testconfig;
 
 import com.coupop.fcfscoupon.domain.fcfs.model.FcfsIssueRepository;
 import org.springframework.data.redis.core.StringRedisTemplate;
