+++ fcfs-coupon-domain-fcfs/src/test/resources/application.properties
