--- fcfs-issue/src/main/java/com/coupop/fcfsissue/exception/CouponNotOpenedException.java
@@ -1,16 +1,0 @@
-package com.coupop.fcfsissue.exception;
-
-
-import com.coupop.fcfscoupon.common.exception.BadRequestException;
-import com.coupop.fcfsissue.model.FcfsIssuePolicy;
-
-public class CouponNotOpenedException extends BadRequestException {
-
-    private static final String TITLE = "쿠폰이 아직 오픈되지 않았습니다.";
-    private static final String DETAIL = String.format("쿠폰은 %d시 %d분에 오픈됩니다.",
-            FcfsIssuePolicy.getOpenAt().getHour(), FcfsIssuePolicy.getOpenAt().getMinute());
-
-    public CouponNotOpenedException() {
-        super(TITLE, DETAIL);
-    }
-}
