--- fcfs-issue/src/main/java/com/coupop/fcfsissue/model/FcfsIssueRepository.java
@@ -1,10 +1,0 @@
-package com.coupop.fcfsissue.model;
-
-public interface FcfsIssueRepository {
-
-    Long add(final String email);
-
-    Long remove(final String email);
-
-    Long getCount();
-}
