--- fcfs-issue/src/test/java/com/coupop/TestContext.java
@@ -1,7 +1,0 @@
-package com.coupop;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TestContext {
-}
