+++ settings.gradle.kts
@@ -1,4 +1,4 @@
 rootProject.name = "fcfs-coupon"
-include("fcfs-issue", "api")
+include("fcfs-coupon-domain-fcfs", "api")
 include("fcfs-coupon-domain-coupon")
 include("fcfs-coupon-common")
